+++ modules/client/src/addresslistVoting/internal/utils.ts
@@ -166,19 +166,38 @@ export function computeProposalStatus(
     parseInt(proposal.startDate) * 1000,
   );
   const endDate = new Date(parseInt(proposal.endDate) * 1000);
+  // The proposal is executed so the status becomes EXECUTED
+  // independently of the other conditions
   if (proposal.executed) {
     return ProposalStatus.EXECUTED;
   }
+  // The proposal is not executed and the start date is in the future
+  // so the status becomes PENDING
   if (startDate >= now) {
     return ProposalStatus.PENDING;
   }
-  if (proposal.potentiallyExecutable || proposal.earlyExecutable) {
+  // The proposal is not executed and the start date is in the past.
+  // Accordingly, we check if the proposal reached enough approval 
+  // (i.e., that the supportThreshold and minParticipation criteria are both met).
+  // If the approvalReached = true and the vote has ended (end date is in the past) it has succeeded. 
+  // This applies to normal mode and vote replacement mode.
+  if (proposal.approvalReached && endDate <= now) {
     return ProposalStatus.SUCCEEDED;
   }
-  if (endDate >= now) {
+  // In early exeuction mode, we calculate if subsequent voting can change the result of the vote. 
+  // If not, the proposal is early executable and is therefore succeeded as well.
+  if(proposal.earlyExecutable){
+    return ProposalStatus.SUCCEEDED;
+  }
+  // The proposal is not executed and the start date is in the past
+  // and the approval threshold is not reached
+  // If the end date is in the future this means that you can still vote
+  // so the status becomes ACTIVE
+  if (now < endDate) {
     return ProposalStatus.ACTIVE;
   }
-  return ProposalStatus.DEFEATED;
+  // If none of the other conditions are met the status becomes DEFEATED
+  return ProposalStatus.DEFEATED
 }
 
 export function computeProposalStatusFilter(status: ProposalStatus) {
@@ -195,7 +214,12 @@ export function computeProposalStatusFilter(status: ProposalStatus) {
       where = { executed: true };
       break;
     case ProposalStatus.SUCCEEDED:
-      where = { potentiallyExecutable: true, endDate_lt: now };
+      where = {
+        or: [
+          { approvalReached: true, endDate_lt: now },
+          { earlyExecutable: true },
+        ],
+      };
       break;
     case ProposalStatus.DEFEATED:
       where = {