+++ modules/client/test/unit/addresslistVoting-client/utils.test.ts
@@ -11,7 +11,7 @@ describe("addresslistVoting-client utils", () => {
       expect(computeProposalStatus({
         endDate: endDate.toString(),
         startDate: startDate.toString(),
-        potentiallyExecutable: false,
+        approvalReached: false,
         executed: false,
         earlyExecutable: false,
       } as SubgraphAddresslistVotingProposal)).toBe(ProposalStatus.PENDING);
@@ -23,7 +23,7 @@ describe("addresslistVoting-client utils", () => {
       expect(computeProposalStatus({
         endDate: endDate.toString(),
         startDate: startDate.toString(),
-        potentiallyExecutable: false,
+        approvalReached: false,
         executed: true,
         earlyExecutable: false,
       } as SubgraphAddresslistVotingProposal)).toBe(ProposalStatus.EXECUTED);
@@ -35,33 +35,72 @@ describe("addresslistVoting-client utils", () => {
       expect(computeProposalStatus({
         endDate: endDate.toString(),
         startDate: startDate.toString(),
-        potentiallyExecutable: false,
+        approvalReached: false,
         executed: false,
         earlyExecutable: false,
       } as SubgraphAddresslistVotingProposal)).toBe(ProposalStatus.ACTIVE);
     });
-    it("should return SUCCEDED if executable = true", () => {
+    it("should return SUCCEEDED if executable = true", () => {
       const endDate = Date.now() / 1000;
       const startDate = (Date.now() / 1000) - 500;
 
       expect(computeProposalStatus({
         endDate: endDate.toString(),
         startDate: startDate.toString(),
-        potentiallyExecutable: true,
+        approvalReached: true,
         executed: false,
         earlyExecutable: false,
       } as SubgraphAddresslistVotingProposal)).toBe(ProposalStatus.SUCCEEDED);
     });
-    it("should return SUCCEDED if earlyExecutable = true", () => {
+    it("should return SUCCEEDED if earlyExecutable = true", () => {
       const endDate = (Date.now() / 1000) + 500;
       const startDate = (Date.now() / 1000) - 500;
 
       expect(computeProposalStatus({
         endDate: endDate.toString(),
         startDate: startDate.toString(),
-        potentiallyExecutable: false,
+        approvalReached: true,
         executed: false,
         earlyExecutable: true,
+      } as SubgraphAddresslistVotingProposal)).toBe(ProposalStatus.SUCCEEDED);
+    });
+    it("should return SUCCEEDED if earlyExecutable = true and signaling proposal", () => {
+      const endDate = (Date.now() / 1000) + 500;
+      const startDate = (Date.now() / 1000) - 500;
+
+      expect(computeProposalStatus({
+        endDate: endDate.toString(),
+        startDate: startDate.toString(),
+        approvalReached: true,
+        executed: false,
+        isSignaling: false,
+        earlyExecutable: true,
+      } as SubgraphAddresslistVotingProposal)).toBe(ProposalStatus.SUCCEEDED);
+    });
+    it("should return SUCCEEDED if approvalReached = true and endDate has passed", () => {
+      const endDate = (Date.now() / 1000) - 200;
+      const startDate = (Date.now() / 1000) - 500;
+
+      expect(computeProposalStatus({
+        endDate: endDate.toString(),
+        startDate: startDate.toString(),
+        approvalReached: true,
+        executed: false,
+        isSignaling: false,
+        earlyExecutable: false,
+      } as SubgraphAddresslistVotingProposal)).toBe(ProposalStatus.SUCCEEDED);
+    });
+    it("should return SUCCEEDED if approvalReached = true and endDate has passed and isSignaling = true", () => {
+      const endDate = (Date.now() / 1000) - 200;
+      const startDate = (Date.now() / 1000) - 500;
+
+      expect(computeProposalStatus({
+        endDate: endDate.toString(),
+        startDate: startDate.toString(),
+        approvalReached: true,
+        executed: false,
+        isSignaling: true,
+        earlyExecutable: false,
       } as SubgraphAddresslistVotingProposal)).toBe(ProposalStatus.SUCCEEDED);
     });
     it("should return DEFEATED", () => {
@@ -71,7 +110,7 @@ describe("addresslistVoting-client utils", () => {
       expect(computeProposalStatus({
         endDate: endDate.toString(),
         startDate: startDate.toString(),
-        potentiallyExecutable: false,
+        approvalReached: false,
         executed: false,
         earlyExecutable: false,
       } as SubgraphAddresslistVotingProposal)).toBe(ProposalStatus.DEFEATED);