+++ modules/client/test/unit/tokenVoting-client  /utils.test.ts
@@ -83,7 +83,7 @@ describe("tokenVoting-client utils", () => {
       expect(computeProposalStatus({
         endDate: endDate.toString(),
         startDate: startDate.toString(),
-        potentiallyExecutable: false,
+        approvalReached: false,
         executed: false,
         earlyExecutable: false,
       } as SubgraphTokenVotingProposal)).toBe(ProposalStatus.PENDING);
@@ -95,7 +95,7 @@ describe("tokenVoting-client utils", () => {
       expect(computeProposalStatus({
         endDate: endDate.toString(),
         startDate: startDate.toString(),
-        potentiallyExecutable: false,
+        approvalReached: false,
         executed: true,
         earlyExecutable: false,
       } as SubgraphTokenVotingProposal)).toBe(ProposalStatus.EXECUTED);
@@ -107,33 +107,72 @@ describe("tokenVoting-client utils", () => {
       expect(computeProposalStatus({
         endDate: endDate.toString(),
         startDate: startDate.toString(),
-        potentiallyExecutable: false,
+        approvalReached: false,
         executed: false,
         earlyExecutable: false,
       } as SubgraphTokenVotingProposal)).toBe(ProposalStatus.ACTIVE);
     });
-    it("should return SUCCEDED if executable = true", () => {
+    it("should return SUCCEEDED if executable = true", () => {
       const endDate = Date.now() / 1000;
       const startDate = (Date.now() / 1000) - 500;
 
       expect(computeProposalStatus({
         endDate: endDate.toString(),
         startDate: startDate.toString(),
-        potentiallyExecutable: true,
+        approvalReached: true,
         executed: false,
         earlyExecutable: false,
       } as SubgraphTokenVotingProposal)).toBe(ProposalStatus.SUCCEEDED);
     });
-    it("should return SUCCEDED if earlyExecutable = true", () => {
+    it("should return SUCCEEDED if earlyExecutable = true", () => {
       const endDate = (Date.now() / 1000) + 500;
       const startDate = (Date.now() / 1000) - 500;
 
       expect(computeProposalStatus({
         endDate: endDate.toString(),
         startDate: startDate.toString(),
-        potentiallyExecutable: false,
+        approvalReached: true,
         executed: false,
         earlyExecutable: true,
+      } as SubgraphTokenVotingProposal)).toBe(ProposalStatus.SUCCEEDED);
+    });
+    it("should return SUCCEEDED if earlyExecutable = true and isSignaling = true", () => {
+      const endDate = (Date.now() / 1000) + 500;
+      const startDate = (Date.now() / 1000) - 500;
+
+      expect(computeProposalStatus({
+        endDate: endDate.toString(),
+        startDate: startDate.toString(),
+        approvalReached: true,
+        executed: false,
+        isSignaling: false,
+        earlyExecutable: true,
+      } as SubgraphTokenVotingProposal)).toBe(ProposalStatus.SUCCEEDED);
+    });
+    it("should return SUCCEEDED if approvalReached = true and endDate has passed", () => {
+      const endDate = (Date.now() / 1000) - 200;
+      const startDate = (Date.now() / 1000) - 500;
+
+      expect(computeProposalStatus({
+        endDate: endDate.toString(),
+        startDate: startDate.toString(),
+        approvalReached: true,
+        executed: false,
+        isSignaling: false,
+        earlyExecutable: false,
+      } as SubgraphTokenVotingProposal)).toBe(ProposalStatus.SUCCEEDED);
+    });
+    it("should return SUCCEEDED if approvalReached = true and endDate has passed and isSignaling = true", () => {
+      const endDate = (Date.now() / 1000) - 200;
+      const startDate = (Date.now() / 1000) - 500;
+
+      expect(computeProposalStatus({
+        endDate: endDate.toString(),
+        startDate: startDate.toString(),
+        approvalReached: true,
+        executed: false,
+        isSignaling: true,
+        earlyExecutable: false,
       } as SubgraphTokenVotingProposal)).toBe(ProposalStatus.SUCCEEDED);
     });
     it("should return DEFEATED", () => {
@@ -143,7 +182,7 @@ describe("tokenVoting-client utils", () => {
       expect(computeProposalStatus({
         endDate: endDate.toString(),
         startDate: startDate.toString(),
-        potentiallyExecutable: false,
+        approvalReached: false,
         executed: false,
         earlyExecutable: false,
       } as SubgraphTokenVotingProposal)).toBe(ProposalStatus.DEFEATED);