+++ modules/client/CHANGELOG.md
@@ -26,7 +26,12 @@ TEMPLATE:
 ### Fixed
 - Fixed proposal id not being transformed
 
+
 ## [1.19.1]
+
+### Added
+
+- Classify signalling multisig proposals as succeeded even after the end date is passed using the subgraph `isSignalling` attribute.  
 
 ### Changed
 
@@ -43,8 +48,6 @@ TEMPLATE:
 - Rename `isPluginUpdateValid` to `isPluginUpdateProposalValid` 
 - Rename `isPluginUpdateAction` to `containsPluginUpdateAction`
 - Rename `isPluginUpdateActionBlockWithRootPermission` to `containsPluginUpdateActionBlockWithRootPermission`
-
-
 
 ## [1.19.0]